requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
aspose-words==22.11.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.0.0
captcha==0.4
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.5
conda==22.9.0
conda-package-handling @ file:///croot/conda-package-handling_1666940373510/work
cryptography @ file:///croot/cryptography_1665612644927/work
cssselect==1.1.0
-e git+https://github.com/gfjiangly/cvtools.git@071a157ecd9bed4c031244861db91dd212050245#egg=cvtoolss
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dicomsdl==0.109.1
docxcompose==1.0.0a9
-e git+https://github.com/giuse88/duka.git@76415dc6ea41141096c4ca4c50196cc6c077ac30#egg=duka
entrypoints==0.4
executing==0.8.3
fake-useragent==0.1.13
fastapi==0.87.0
fastjsonschema==2.16.1
Flask==2.2.2
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.0
idna @ file:///croot/idna_1666125576474/work
imagesize==1.4.1
importlib-metadata==4.11.3
ipykernel==6.15.0
ipython==8.3.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.9
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
kaggle==1.5.12
kiwisolver==1.4.2
litereval==0.0.11
lxml==4.9.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mmcv==1.6.2
multidict==6.0.2
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
nvidia-dali-cuda110==1.20.0
nvidia-dali-nightly-cuda110==1.22.0.dev20221215
oauthlib==3.2.0
onnxruntime-gpu==1.11.1
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.2
pydicom==2.3.1
pyee==8.2.2
pygdcm==1.0
Pygments==2.12.0
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.8
pyppdf==0.1.2
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-docx==0.8.6
python-gdcm==3.0.20
python-slugify==7.0.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.8
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
scikit-learn==1.1.3
scipy==1.8.0
seaborn==0.11.2
-e git+https://github.com/bisohns/search-engine-parser.git@0c2f4bde7dd21c10e64c9204417d9a228e96c187#egg=search_engine_parser
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
starlette==0.21.0
system-health-publisher @ file:///mnt/hdd2T/vietanhdev/system_health_publisher
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
terminaltables==3.1.10
text-unidecode==1.3
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tinycss2==1.1.1
toolz @ file:///croot/toolz_1667464077321/work
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.3.0
typing_extensions==4.2.0
urllib3 @ file:///croot/urllib3_1666298941550/work
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.2.2
wget==3.2
yapf==0.32.0
yarl==1.8.1
zipp==3.8.0
